HW2: correction mistakes

+++ src/task_4_8.py
@@ -3,7 +3,7 @@
 число. Возвести его в куб.
 """
 
-num = int(input('Enter the number: '))
+num = float(input('Enter the number: '))
 
 if type(num) == int or float:
     print(num ** 3)
